+++ JackedEngine/Scene/Actors/VikingRoomActor.h
@@ -6,7 +6,7 @@ class VikingRoomActor : public BaseActor{
 public:
 	VikingRoomActor(ActorInitializer initializer);
 	
-	virtual void Tick(float deltaTime) override;
+	virtual void Tick(double deltaTime) override;
 
 private:
 	GenericMeshMaterial meshMaterial;
